refactor: Move utilities into a commonly shared place
