+++ ors-benchmark/src/main/java/org/heigit/ors/benchmark/IsochronesLoadTest.java
@@ -1,30 +1,25 @@
 package org.heigit.ors.benchmark;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import io.gatling.javaapi.core.PopulationBuilder;
+import io.gatling.javaapi.core.ScenarioBuilder;
+import io.gatling.javaapi.core.Session;
+import io.gatling.javaapi.http.HttpRequestActionBuilder;
+import org.heigit.ors.benchmark.BenchmarkEnums.RangeType;
+import org.heigit.ors.exceptions.RequestBodyCreationException;
+import org.heigit.ors.util.SourceUtils;
+import org.slf4j.LoggerFactory;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Stream;
 
-import org.heigit.ors.benchmark.BenchmarkEnums.RangeType;
-import static org.heigit.ors.benchmark.util.NameUtils.getFileNameWithoutExtension;
-import org.heigit.ors.benchmark.util.SourceUtils;
-import org.heigit.ors.exceptions.RequestBodyCreationException;
-import org.slf4j.LoggerFactory;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-
-import static io.gatling.javaapi.core.CoreDsl.StringBody;
-import static io.gatling.javaapi.core.CoreDsl.atOnceUsers;
-import static io.gatling.javaapi.core.CoreDsl.csv;
-import static io.gatling.javaapi.core.CoreDsl.feed;
-import static io.gatling.javaapi.core.CoreDsl.scenario;
-import io.gatling.javaapi.core.PopulationBuilder;
-import io.gatling.javaapi.core.ScenarioBuilder;
-import io.gatling.javaapi.core.Session;
+import static io.gatling.javaapi.core.CoreDsl.*;
 import static io.gatling.javaapi.http.HttpDsl.http;
 import static io.gatling.javaapi.http.HttpDsl.status;
-import io.gatling.javaapi.http.HttpRequestActionBuilder;
+import static org.heigit.ors.util.NameUtils.getFileNameWithoutExtension;
 
 public class IsochronesLoadTest extends AbstractLoadTest {
