+++ ors-benchmark/src/test/java/org/heigit/ors/util/NameUtilsTest.java
@@ -1,13 +1,14 @@
-package org.heigit.ors.benchmark.util;
+package org.heigit.ors.util;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
 class NameUtilsTest {
